Enum google_oauth2_v2::Scope

pub enum Scope {
    PluLogin,
    UserinfoEmail,
    UserinfoProfile,
    PluMe,
}

Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Variants

PluLogin

Know your basic profile info and list of people in your circles.

UserinfoEmail

View your email address

UserinfoProfile

View your basic profile info

PluMe

Know who you are on Google