Struct google_oauth2_v2::Userinfoplus

pub struct Userinfoplus {
    pub family_name: Option<String>,
    pub name: Option<String>,
    pub picture: Option<String>,
    pub locale: Option<String>,
    pub gender: Option<String>,
    pub id: Option<String>,
    pub link: Option<String>,
    pub given_name: Option<String>,
    pub email: Option<String>,
    pub hd: Option<String>,
    pub verified_email: Option<bool>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

v2 me get userinfo (response)

get userinfo (response)

Fields

family_name

The user's last name.

name

The user's full name.

picture

URL of the user's picture image.

locale

The user's preferred locale.

gender

The user's gender.

id

The obfuscated ID of the user.

link

URL of the profile page.

given_name

The user's first name.

email

The user's email address.

hd

The hosted domain e.g. example.com if the user is Google apps user.

verified_email

Boolean flag which is true if the email address is verified. Always verified because we only return the user's primary email address.